Moved test decks to app file
