+++ packages/webapp/src/pages/collection.tsx
@@ -27,7 +27,7 @@ const initialEffectMap = Object.assign({}, ...effects.map(name => ({[name]: fals
 const types = ['Creature', 'Magic', 'Weapon']
 const initialTypeMap = Object.assign({}, ...types.map(name => ({[name]: false})))
 
-const Collection: FablePage = ({ isHydrated }) => {
+const Collection: FablePage = ({ decks, isHydrated }) => {
 
   const { address } = useAccount()
   const [ selectedCard, setSelectedCard ] = useState<Card|null>(null)
@@ -75,60 +75,6 @@ const Collection: FablePage = ({ isHydrated }) => {
     const type = types[typeIndex]
     setTypeMap({...typeMap, [type]: !typeMap[type]})
   }
-
-  // todo @eviterin: would be good to have something in read.ts that allows me to fetch all decks by address
-  const testCards = [
-    {
-      id: BigInt(57),
-      lore: {
-        name: "Fire Fighter",
-        flavor: "",
-        URL: ""
-      },
-      stats: {
-        attack: 2,
-        defense: 2
-      },
-      cardTypeID: 57
-    },
-    {
-      id: BigInt(31),
-      lore: {
-        name: "Wise Elf",
-        flavor: "",
-        URL: ""
-      },
-      stats: {
-        attack: 1,
-        defense: 3
-      },
-      cardTypeID: 31
-    },
-    {
-      id: BigInt(38),
-      lore: {
-        name: "Grave Digger",
-        flavor: "",
-        URL: ""
-      },
-      stats: {
-        attack: 2,
-        defense: 3
-      },
-      cardTypeID: 38
-    }
-  ];
-  const [decks, setDecks] = useState([
-    {
-      name: "Mystical Creatures",
-      cards: testCards // Assuming 'cards' contains the array of card objects
-    },
-    {
-      name: "Empty Deck",
-      cards: [] // An empty deck for testing
-    }
-  ]);
-
 
   return (
     <>
